Declaration filler test
